requirements.txt
Flask==3.1.1
python-dotenv==1.1.1
pymongo==4.13.2
Werkzeug==3.1.3
requests==2.32.4
yfinance==0.2.65
pandas==2.3.1
pytest==8.4.1
mongomock==4.3.0gunicorn==23.0.0
